requirements.txt
django
djangorestframework
gunicorn
django-heroku
djangorestframework_simplejwt
djoser
whitenoise
xlrd

